requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
APScheduler>=3.10.4
requests>=2.31.0
parsel>=1.8.1
lxml>=4.9.3
beautifulsoup4>=4.12.0
selenium>=4.15.0  
PyMySQL>=1.1.0
cryptography>=41.0.7
python-dotenv>=1.0.0
psycopg2-binary  
oracledb         
